# src/modules/users/controllers/users.controller.ts
import {
  Controller,
  Get,
  Post,
  Patch,
  Delete,
  Body,
  Param,
  Query,
  HttpCode,
  HttpStatus,
  UploadedFile,
  BadRequestException,
} from '@nestjs/common'
import {
  ApiTags,
  ApiOperation,
  ApiResponse,
  ApiBearerAuth,
} from '@nestjs/swagger'
import {
  ApiCreate,
  ApiList,
  ApiFindOne,
  ApiOkResponse,
  ApiNotFoundResponse,
  ApiStandardResponses,
  ApiUpdateWithMessage,
  ApiRemoveWithMessage,
} from '@core/swagger'
import { UuidParamDto } from '@core/dtos'
import { ResponseMessage } from '@core/http'

import {
  CreateUserUseCase,
  CreateUserDto,
  UpdateUserUseCase,
  UpdateUserDto,
  FindAllUsersUseCase,
  FindUserByIdUseCase,
  UploadProfileImageUseCase,
  DeleteProfileImageUseCase,
  DeactivateUserUseCase,
  RemoveUserUseCase,
  ActivateUserUseCase,
  TwoFactorActivateUserUseCase,
  TwoFactorDeactivateUserUseCase,
} from '../use-cases'
import {
  FindUsersDto,
  USER_SORTABLE_FIELDS,
  USER_SEARCH_FIELDS,
  UserResponseDto,
} from '../dtos'
import { UploadAvatar } from '@core/files'
import { Role } from '@core'

@ApiBearerAuth()
@ApiTags('users')
@Controller('users')
export class UsersController {
  constructor(
    private readonly createUserUseCase: CreateUserUseCase,
    private readonly updateUserUseCase: UpdateUserUseCase,
    private readonly findAllUsersUseCase: FindAllUsersUseCase,
    private readonly findUserByIdUseCase: FindUserByIdUseCase,
    private readonly uploadProfileImageUseCase: UploadProfileImageUseCase,
    private readonly deleteProfileImageUseCase: DeleteProfileImageUseCase,
    private readonly removeUserUseCase: RemoveUserUseCase,
    private readonly activateUserUseCase: ActivateUserUseCase,
    private readonly deactivateUserUseCase: DeactivateUserUseCase,
    private readonly twoFactorActivateUserUseCase: TwoFactorActivateUserUseCase,
    private readonly twoFactorDeactivateUserUseCase: TwoFactorDeactivateUserUseCase,
  ) {}

  @Post()
  @ApiCreate(UserResponseDto, {
    summary: 'Crear un nuevo usuario',
    description:
      'Crea un nuevo usuario con sus datos básicos. La contraseña se hashea automáticamente con bcrypt.',
    conflictMessage: 'Ya existe un usuario con ese email, username o CI',
  })
  async create(@Body() createUserDto: CreateUserDto) {
    return await this.createUserUseCase.execute(createUserDto)
  }

  @Get()
  @ApiList(UserResponseDto, {
    summary: 'Listar todos los usuarios',
    searchFields: USER_SEARCH_FIELDS,
    sortableFields: USER_SORTABLE_FIELDS.map(String),
    defaultSortBy: 'createdAt',
    filterFields: [
      {
        name: 'role',
        description: 'Filtrar por rol',
        type: `enum: ${Object.values(Role).join(', ')}`,
      },
      {
        name: 'organizationId',
        description: 'Filtrar por ID de organización',
        type: 'UUID',
      },
    ],
  })
  async findAll(@Query() findUsersDto: FindUsersDto) {
    return await this.findAllUsersUseCase.execute(findUsersDto)
  }

  @Get(':id')
  @ApiFindOne(UserResponseDto)
  async findOne(@Param() { id }: UuidParamDto) {
    return await this.findUserByIdUseCase.execute(id)
  }

  // OPCIÓN 1: Devolver entidad actualizada (RECOMENDADO para frontends modernos)
  // @Patch(':id')
  // @ApiUpdate(UserResponseDto, {
  //   description:
  //     'Actualiza los datos de un usuario y retorna la entidad actualizada. NO actualiza la contraseña (usar endpoint de autenticación).',
  //   conflictMessage: 'Ya existe un usuario con ese email, username o CI',
  // })
  // async update(
  //   @Param() { id }: UuidParamDto,
  //   @Body() updateUserDto: UpdateUserDto,
  // ) {
  //   return await this.updateUserUseCase.execute(id, updateUserDto)
  // }

  // OPCIÓN 2: Devolver mensaje genérico (más ligero)
  @Patch(':id')
  @ResponseMessage('Usuario actualizado exitosamente')
  @ApiUpdateWithMessage({
    summary: 'Actualizar usuario',
    description:
      'Actualiza los datos de un usuario y retorna un mensaje de confirmación. NO actualiza la contraseña (usar endpoint de autenticación).',
    conflictMessage: 'Ya existe un usuario con ese email, username o CI',
  })
  async update(
    @Param() { id }: UuidParamDto,
    @Body() updateUserDto: UpdateUserDto,
  ) {
    await this.updateUserUseCase.execute(id, updateUserDto)
  }

  @Post(':id/upload-image')
  @HttpCode(HttpStatus.OK)
  @UploadAvatar({
    maxSize: 2 * 1024 * 1024, // 2MB
  })
  @ApiOperation({
    summary: 'Subir imagen de perfil del usuario',
    description:
      'Sube o reemplaza la imagen de perfil y retorna el usuario actualizado. Formatos: JPG, PNG, WebP. Tamaño máximo: 2MB. Dimensiones: 512x512px.',
  })
  @ApiOkResponse(UserResponseDto, 'Imagen subida exitosamente')
  @ApiNotFoundResponse('Usuario no encontrado')
  @ApiStandardResponses()
  async uploadProfileImage(
    @Param() { id }: UuidParamDto,
    @UploadedFile() file: Express.Multer.File,
  ) {
    if (!file) {
      throw new BadRequestException('Debe proporcionar un archivo de imagen')
    }
    return await this.uploadProfileImageUseCase.execute(id, file)
  }

  @Delete(':id/image')
  @HttpCode(HttpStatus.OK)
  @ApiOperation({
    summary: 'Eliminar imagen de perfil del usuario',
    description:
      'Elimina la imagen de perfil del usuario, la remueve del storage y retorna el usuario actualizado.',
  })
  @ApiOkResponse(UserResponseDto, 'Imagen eliminada exitosamente')
  @ApiNotFoundResponse('Usuario no encontrado')
  @ApiStandardResponses()
  async deleteProfileImage(@Param() { id }: UuidParamDto) {
    return await this.deleteProfileImageUseCase.execute(id)
  }

  // OPCIÓN 1: Devolver entidad actualizada
  // @Patch(':id/deactivate')
  // @ApiCustom(UserResponseDto, {
  //   summary: 'Desactivar un usuario',
  //   description:
  //     'Cambia el estado del usuario a SUSPENDED y retorna el usuario actualizado.',
  // })
  // async deactivate(@Param() { id }: UuidParamDto) {
  //   return await this.deactivateUserUseCase.execute(id)
  // }

  // OPCIÓN 2: Devolver mensaje genérico (usa TransformInterceptor + @ResponseMessage)
  @Patch(':id/deactivate')
  @ResponseMessage('Usuario desactivado exitosamente')
  @ApiUpdateWithMessage({
    summary: 'Desactivar un usuario',
    description:
      'Cambia el estado del usuario a SUSPENDED y retorna un mensaje de confirmación.',
  })
  async deactivate(@Param() { id }: UuidParamDto) {
    await this.deactivateUserUseCase.execute(id)
  }

  @Patch(':id/activate')
  @ResponseMessage('Usuario activado exitosamente')
  @ApiUpdateWithMessage({
    summary: 'Activar un usuario',
    description:
      'Cambia el estado del usuario a ACTIVE y retorna un mensaje de confirmación.',
  })
  async activate(@Param() { id }: UuidParamDto) {
    await this.activateUserUseCase.execute(id)
  }

  @Patch(':id/2fa/activate')
  @ResponseMessage('two factor activado exitosamente')
  @ApiUpdateWithMessage({
    summary: 'Activar 2FA de un usuario',
    description:
      'Cambia el estado del 2fa de un usuario a ACTIVE y retorna un mensaje de confirmación.',
  })
  async activate2fa(@Param() { id }: UuidParamDto) {
    await this.twoFactorActivateUserUseCase.execute(id)
  }

  @Patch(':id/2fa/deactivate')
  @ResponseMessage('two factor activado exitosamente')
  @ApiUpdateWithMessage({
    summary: 'Desactivar 2fa un usuario',
    description:
      'Cambia el estado del 2fa de un usuario a INACTIVE y retorna un mensaje de confirmación.',
  })
  async deactivate2fa(@Param() { id }: UuidParamDto) {
    await this.twoFactorDeactivateUserUseCase.execute(id)
  }

  // OPCIÓN 1: Devolver entidad eliminada (útil para confirmación visual)
  // @Delete(':id')
  // @ApiRemove(UserResponseDto, {
  //   summary: 'Eliminar un usuario (soft delete)',
  //   description:
  //     'Marca el usuario como eliminado sin borrar sus datos de la base de datos. Retorna el usuario eliminado para confirmación.',
  // })
  // async remove(@Param() { id }: UuidParamDto) {
  //   return await this.removeUserUseCase.execute(id)
  // }

  // OPCIÓN 2: Devolver mensaje genérico
  @Delete(':id')
  @ResponseMessage('Usuario eliminado exitosamente')
  @ApiRemoveWithMessage({
    summary: 'Eliminar un usuario (soft delete)',
    description:
      'Marca el usuario como eliminado sin borrar sus datos de la base de datos. Retorna un mensaje de confirmación.',
  })
  async remove(@Param() { id }: UuidParamDto) {
    await this.removeUserUseCase.execute(id)
  }

  // OPCIÓN 3: 204 No Content (REST puro - sin cuerpo de respuesta)
  // @Delete(':id')
  // @ApiRemoveNoContent({
  //   summary: 'Eliminar un usuario (soft delete)',
  //   description:
  //     'Marca el usuario como eliminado sin borrar sus datos de la base de datos. No devuelve contenido.',
  // })
  // async remove(@Param() { id }: UuidParamDto) {
  //   await this.removeUserUseCase.execute(id)
  //   // No devolver nada (NestJS automáticamente envía 204)
  // }
}