# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Crea un nuevo usuario con sus datos básicos. La contraseña se hashea automáticamente con bcrypt.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de página (empieza en 1)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Cantidad de registros por página",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "all",
            "required": false,
            "in": "query",
            "description": "Si es true, devuelve TODOS los registros (ignora page y limit)",
            "schema": {
              "default": false,
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Campo por el cual ordenar (Solo uno)",
            "schema": {
              "example": "createdAt",
              "type": "string",
              "enum": [
                "lastNames",
                "email",
                "createdAt",
                "isActive",
                "ci",
                "phone",
                "names",
                "updatedAt"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Dirección del ordenamiento (ascendente o descendente)",
            "schema": {
              "default": "DESC",
              "example": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filtrar por estado activo/inactivo",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "description": "Filtrar usuarios por ID de organización (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de recursos obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recurso encontrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Actualiza los datos de un usuario y retorna un mensaje de confirmación. NO actualiza la contraseña (usar endpoint de autenticación).",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marca el usuario como eliminado sin borrar sus datos de la base de datos. Retorna un mensaje de confirmación.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/upload-image": {
      "post": {
        "description": "Sube o reemplaza la imagen de perfil y retorna el usuario actualizado. Formatos: JPG, PNG, WebP. Tamaño máximo: 2MB. Dimensiones: 512x512px.",
        "operationId": "UsersController_uploadProfileImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Subir avatar de usuario (JPG, PNG, WebP)",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary",
                    "description": "Archivo (.jpg, .jpeg, .png, .gif, .webp, .svg)"
                  }
                },
                "required": [
                  "avatar"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imagen subida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Subir imagen de perfil del usuario",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/image": {
      "delete": {
        "description": "Elimina la imagen de perfil del usuario, la remueve del storage y retorna el usuario actualizado.",
        "operationId": "UsersController_deleteProfileImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Imagen eliminada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Eliminar imagen de perfil del usuario",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/deactivate": {
      "patch": {
        "operationId": "UsersController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cambia el estado del usuario a SUSPENDED y retorna un mensaje de confirmación.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/activate": {
      "patch": {
        "operationId": "UsersController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cambia el estado del usuario a ACTIVE y retorna un mensaje de confirmación.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/2fa/activate": {
      "patch": {
        "operationId": "UsersController_activate2fa",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cambia el estado del 2fa de un usuario a ACTIVE y retorna un mensaje de confirmación.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/2fa/deactivate": {
      "patch": {
        "operationId": "UsersController_deactivate2fa",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cambia el estado del 2fa de un usuario a INACTIVE y retorna un mensaje de confirmación.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "ATR API",
    "description": "API de auditorías, plantillas, frameworks de madurez y gestión de usuarios",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indica si la operación fue exitosa",
            "example": true,
            "default": true
          },
          "statusCode": {
            "type": "number",
            "description": "Código de estado HTTP",
            "example": 200,
            "enum": [
              200,
              201,
              204
            ]
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo de la operación. Se genera automáticamente según el método HTTP o se personaliza con @ResponseMessage",
            "example": "Operación exitosa"
          },
          "data": {
            "type": "object",
            "description": "Datos de la operación. Será null si el endpoint no retorna datos.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp ISO 8601 de cuando se generó la respuesta",
            "example": "2024-01-30T10:30:00.000Z"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "timestamp"
        ]
      },
...
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo de la operación",
            "example": "Operación completada exitosamente"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "names": {
            "type": "string",
            "description": "Nombres del usuario",
            "example": "Juan Carlos",
            "minLength": 2,
            "maxLength": 50
          },
          "lastNames": {
            "type": "string",
            "description": "Apellidos del usuario",
            "example": "Pérez García",
            "minLength": 2,
            "maxLength": 50
          },
          "email": {
            "type": "string",
            "description": "Email del usuario",
            "example": "juan.perez@example.com",
            "maxLength": 100
          },
          "username": {
            "type": "string",
            "description": "Nombre de usuario (solo letras, números y guión bajo)",
            "example": "juan_perez",
            "minLength": 3,
            "maxLength": 30
          },
          "ci": {
            "type": "string",
            "description": "Cédula de Identidad",
            "example": "1234567-8",
            "minLength": 5,
            "maxLength": 15
          },
          "phone": {
            "type": "string",
            "description": "Teléfono de contacto",
            "example": "+591 2 234 5678",
            "maxLength": 20
          },
          "address": {
            "type": "string",
            "description": "Dirección del usuario",
            "example": "Av. 6 de Agosto #123, La Paz",
            "maxLength": 200
          },
          "organizationId": {
            "type": "string",
            "description": "ID de la organización a la que pertenece el usuario (requerido)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "roles": {
            "type": "array",
            "description": "Roles del usuario",
            "example": [
              "auditor"
            ],
            "items": {
              "type": "string",
              "enum": [
                "admin",
                "gerente",
                "auditor",
                "cliente"
              ]
            }
          }
        },
        "required": [
          "names",
          "lastNames",
          "email",
          "username",
          "ci",
          "organizationId",
          "roles"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "names": {
            "type": "string",
            "description": "Nombres del usuario",
            "example": "Juan Carlos",
            "minLength": 2,
            "maxLength": 50
          },
          "lastNames": {
            "type": "string",
            "description": "Apellidos del usuario",
            "example": "Pérez García",
            "minLength": 2,
            "maxLength": 50
          },
          "username": {
            "type": "string",
            "description": "Nombre de usuario (solo letras, números y guión bajo)",
            "example": "juan_perez",
            "minLength": 3,
            "maxLength": 30
          },
          "ci": {
            "type": "string",
            "description": "Cédula de Identidad",
            "example": "1234567-8",
            "minLength": 5,
            "maxLength": 15
          },
          "phone": {
            "type": "string",
            "description": "Teléfono de contacto",
            "example": "+591 2 234 5678",
            "maxLength": 20
          },
          "address": {
            "type": "string",
            "description": "Dirección del usuario",
            "example": "Av. 6 de Agosto #123, La Paz",
            "maxLength": 200
          },
          "roles": {
            "type": "array",
            "description": "Roles del usuario",
            "example": [
              "auditor"
            ],
            "items": {
              "type": "string",
              "enum": [
                "admin",
                "gerente",
                "auditor",
                "cliente"
              ]
            }
          }
        }
      },
...
    }
  }
}